requirements.txt
asgiref==3.11.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
Django==6.0.4
Flask==3.1.3
greenlet==3.4.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-snap7==3.0.0
requests==2.33.1
SQLAlchemy==2.0.49
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
Werkzeug==3.1.8
